+++ src/pages/documentsPage/DocumentsPage.tsx
@@ -1,58 +1,14 @@
-import {
-    Badge,
-    Box,
-    Button,
-    Center,
-    Flex,
-    FormControl,
-    FormLabel,
-    Grid,
-    GridItem,
-    Input,
-    List,
-    ListItem,
-    Modal,
-    ModalBody,
-    ModalCloseButton,
-    ModalContent,
-    ModalFooter,
-    ModalHeader,
-    ModalOverlay,
-    Select,
-    Spacer,
-    Stack,
-    useDisclosure,
-    useToast,
-    Spinner,
-    Skeleton,
-    Heading,
-    Text,
-    StackDivider,
-    VStack,
-} from "@chakra-ui/react";
+import { Badge, Box, Button, Center, Flex, Grid, GridItem, List, ListItem, useDisclosure } from "@chakra-ui/react";
 import { format } from "date-fns";
-import {
-    arrayUnion,
-    collection,
-    deleteField,
-    doc,
-    getDoc,
-    query,
-    serverTimestamp,
-    setDoc,
-    Timestamp,
-    updateDoc,
-    where,
-} from "firebase/firestore";
-import { auth, db } from "firebaseConfig/firebase";
+import { collection } from "firebase/firestore";
+import { db } from "firebaseConfig/firebase";
 import useDocDataQuery from "hooks/useDocDataQuery";
-import useInput from "hooks/useInput";
-import { useRef, useState, useCallback, useMemo, memo } from "react";
-import DatePicker from "react-datepicker";
+import { useState, useCallback, useMemo } from "react";
 import { useOutletContext } from "react-router-dom";
 import { DocType, UserType } from "types/ts";
-import { getUid } from "utils/common";
 import { useFirestoreQueryData } from "@react-query-firebase/firestore";
+import DocDetailModal from "./DocDetailModal";
+import CreateDocModal from "./CreateDocModal";
 
 const DocumentsPage = () => {
     const docModal = useDisclosure();
@@ -109,7 +65,7 @@ const DocumentsPage = () => {
 };
 
 export default DocumentsPage;
-// { workerDocList, type }: { workerDocList: DocType[] | undefined; type: string }
+
 const DocumentList = ({ type }: { type: string }): JSX.Element => {
     const userInfo: UserType = useOutletContext().userInfo;
     const company = userInfo?.company;
@@ -201,373 +157,7 @@ const DocumentList = ({ type }: { type: string }): JSX.Element => {
                     );
                 }
             })}
-            <DetailModal detailModal={detailModal} clickedData={clickedData} company={company} role={role} />
-        </>
-    );
-};
-
-const CreateDocModal = memo(({ isOpen, onClose }: { isOpen: boolean; onClose: () => void }) => {
-    const [title, setTitle, changeTitle] = useInput("");
-    const [type, setType, changeType] = useInput("");
-    const initialRef = useRef(null);
-    const finalRef = useRef(null);
-    const [startDate, setStartDate] = useState(null);
-    const [endDate, setEndDate] = useState(null);
-    const toast = useToast();
-    const userInfo: UserType = useOutletContext().userInfo;
-    const userName = userInfo?.name;
-    const company = userInfo?.company;
-    const userUid = useOutletContext().userUid;
-
-    const onSubmitHandler = (e: React.FormEvent<HTMLElement>) => {
-        e.preventDefault();
-        if (startDate === null || endDate === null) {
-            toast({
-                title: "기간을 선택해주세요.",
-                status: "error",
-                duration: 5000,
-                isClosable: true,
-            });
-            return;
-        }
-        if (startDate !== null && endDate !== null) {
-            const saveDoc = async () => {
-                const documentUid = getUid();
-                const date = format(new Date(), "yyyy-MM-dd");
-                try {
-                    await setDoc(
-                        doc(db, company, userUid),
-                        {
-                            [documentUid]: {
-                                title,
-                                startDate: format(startDate, "yyyy/MM/dd"),
-                                endDate: format(endDate, "yyyy/MM/dd"),
-                                userUid,
-                                createdAt: serverTimestamp(),
-                                status: "waiting",
-                                documentUid,
-                                type,
-                                name: userName,
-                                date,
-                            },
-                        },
-                        { merge: true }
-                    );
-                    toast({
-                        title: "성공적으로 제출됐습니다.",
-                        status: "success",
-                        duration: 5000,
-                        isClosable: true,
-                    });
-                } catch (error) {
-                    console.log(error);
-                    toast({
-                        title: "제출 실패",
-                        status: "error",
-                        duration: 5000,
-                        isClosable: true,
-                    });
-                }
-            };
-            saveDoc();
-            onClose();
-            setTitle("");
-            setType("");
-            setStartDate(null);
-            setEndDate(null);
-        }
-    };
-    return (
-        <Modal initialFocusRef={initialRef} finalFocusRef={finalRef} isOpen={isOpen} onClose={onClose}>
-            <ModalOverlay />
-            <ModalContent>
-                <Box as="form" onSubmit={onSubmitHandler}>
-                    <ModalHeader>연차 신청서</ModalHeader>
-                    <ModalCloseButton />
-                    <ModalBody pb={6}>
-                        <Stack gap={2}>
-                            <FormControl isRequired>
-                                <FormLabel>제목</FormLabel>
-                                <Input
-                                    type="text"
-                                    value={title}
-                                    onChange={changeTitle}
-                                    placeholder="제목을 입력해주세요."
-                                    variant="flushed"
-                                    _placeholder={{ fontSize: "0.9rem" }}
-                                    ref={initialRef}
-                                />
-                            </FormControl>
-
-                            <FormControl isRequired>
-                                <FormLabel>종류</FormLabel>
-                                <Select
-                                    placeholder="종류를 선택해주세요."
-                                    value={type}
-                                    onChange={changeType}
-                                    variant="flushed"
-                                >
-                                    <option value="full">연차</option>
-                                    <option value="half">반차</option>
-                                </Select>
-                            </FormControl>
-
-                            <FormControl mt={4} isRequired>
-                                <FormLabel>날짜</FormLabel>
-                                <DatePicker
-                                    selected={startDate}
-                                    onChange={(dates: any) => {
-                                        const [start, end] = dates;
-                                        setStartDate(start);
-                                        setEndDate(end);
-                                    }}
-                                    startDate={startDate}
-                                    endDate={endDate}
-                                    selectsRange
-                                    placeholderText="기간을 선택해주세요."
-                                />
-                            </FormControl>
-                        </Stack>
-                    </ModalBody>
-
-                    <ModalFooter>
-                        <Button colorScheme="blue" mr={3} type="submit">
-                            신청
-                        </Button>
-                        <Button onClick={onClose}>취소</Button>
-                    </ModalFooter>
-                </Box>
-            </ModalContent>
-        </Modal>
-    );
-});
-
-const DetailModal = ({
-    detailModal,
-    clickedData,
-    company,
-    role,
-}: {
-    detailModal: {
-        isOpen: boolean;
-        onOpen: () => void;
-        onClose: () => void;
-        onToggle: () => void;
-        isControlled: boolean;
-        getButtonProps: (props?: any) => any;
-        getDisclosureProps: (props?: any) => any;
-    };
-    clickedData: DocType | null;
-    company: string;
-    role: string;
-}) => {
-    const toast = useToast();
-    const initialRef = useRef(null);
-    const rejectModal = useDisclosure();
-    const [rejectReason, setRejectReason, changeRejectReason] = useInput("");
-    const successBtnHandler = useCallback(async () => {
-        console.log(clickedData);
-        try {
-            if (clickedData !== null) {
-                await updateDoc(doc(db, company, clickedData?.userUid), {
-                    [clickedData?.documentUid]: {
-                        createdAt: clickedData?.createdAt,
-                        documentUid: clickedData?.documentUid,
-                        endDate: clickedData?.endDate,
-                        name: clickedData?.name,
-                        startDate: clickedData?.startDate,
-                        status: "success",
-                        title: clickedData?.title,
-                        type: clickedData?.type,
-                        userUid: clickedData?.userUid,
-                        date: clickedData?.date,
-                    },
-                });
-                toast({
-                    title: `결재를 완료하였습니다.`,
-                    status: "success",
-                    duration: 5000,
-                    isClosable: true,
-                });
-                detailModal.onClose();
-            }
-        } catch (error) {
-            console.log(error);
-        }
-    }, [clickedData, toast, company, detailModal]);
-
-    const rejectSubmitHandler = useCallback(
-        async (e: React.FormEvent<HTMLElement>) => {
-            e.preventDefault();
-            try {
-                if (clickedData !== null) {
-                    await updateDoc(doc(db, company, clickedData?.userUid), {
-                        [clickedData?.documentUid]: {
-                            createdAt: clickedData?.createdAt,
-                            documentUid: clickedData?.documentUid,
-                            endDate: clickedData?.endDate,
-                            name: clickedData?.name,
-                            startDate: clickedData?.startDate,
-                            status: "reject",
-                            title: clickedData?.title,
-                            type: clickedData?.type,
-                            userUid: clickedData?.userUid,
-                            date: clickedData?.date,
-                            reject: rejectReason,
-                        },
-                    });
-                    toast({
-                        title: `결재를 완료하였습니다.`,
-                        status: "success",
-                        duration: 5000,
-                        isClosable: true,
-                    });
-                    rejectModal.onClose();
-                    detailModal.onClose();
-                }
-            } catch (error) {
-                console.log(error);
-            }
-        },
-        [clickedData, company, rejectModal, rejectReason, toast, detailModal]
-    );
-
-    const deleteDocHandler = useCallback(async () => {
-        try {
-            if (clickedData !== null) {
-                await updateDoc(doc(db, company, clickedData?.userUid), {
-                    [clickedData?.documentUid]: deleteField(),
-                });
-                detailModal.onClose();
-                toast({
-                    title: `문서 삭제를 성공하였습니다.`,
-                    status: "success",
-                    duration: 5000,
-                    isClosable: true,
-                });
-            }
-        } catch (error) {
-            console.log(error);
-            toast({
-                title: `문서 삭제를 실패하였습니다.`,
-                status: "error",
-                duration: 5000,
-                isClosable: true,
-            });
-        }
-    }, [clickedData, company, detailModal, toast]);
-
-    return (
-        <>
-            <Modal isOpen={detailModal.isOpen} onClose={detailModal.onClose}>
-                <ModalOverlay />
-                <ModalContent>
-                    <ModalHeader>신청서</ModalHeader>
-                    <ModalCloseButton />
-                    <ModalBody>
-                        <VStack divider={<StackDivider bg="gray.200" h="1px" />} spacing="2" align="stretch">
-                            <Box>
-                                <Heading size="xs" textTransform="uppercase" color="gray.500">
-                                    제목
-                                </Heading>
-                                <Text pt="2" fontSize="md">
-                                    {clickedData?.title}
-                                </Text>
-                            </Box>
-                            <Box>
-                                <Heading size="xs" textTransform="uppercase" color="gray.500">
-                                    신청자
-                                </Heading>
-                                <Text pt="2" fontSize="md">
-                                    {clickedData?.name}
-                                </Text>
-                            </Box>
-                            <Box>
-                                <Heading size="xs" textTransform="uppercase" color="gray.500">
-                                    종류
-                                </Heading>
-                                <Text pt="2" fontSize="md">
-                                    {clickedData?.type === "full" ? "연차" : "반차"}
-                                </Text>
-                            </Box>
-                            <Box>
-                                <Heading size="xs" textTransform="uppercase" color="gray.500">
-                                    신청 날짜
-                                </Heading>
-                                <Text pt="2" fontSize="md">
-                                    {`${clickedData?.startDate} ~ ${clickedData?.endDate}`}
-                                </Text>
-                            </Box>
-                            {clickedData?.reject && (
-                                <Box>
-                                    <Heading size="xs" textTransform="uppercase" color="gray.500">
-                                        반려 사유
-                                    </Heading>
-                                    <Text pt="2" fontSize="md">
-                                        {clickedData?.reject}
-                                    </Text>
-                                </Box>
-                            )}
-                        </VStack>
-                    </ModalBody>
-
-                    <ModalFooter>
-                        {clickedData?.status === "waiting" && role === "admin" && (
-                            <>
-                                <Button
-                                    colorScheme="pink"
-                                    mr={3}
-                                    onClick={() => {
-                                        rejectModal.onOpen();
-                                    }}
-                                >
-                                    반려
-                                </Button>
-
-                                <Button colorScheme="blue" onClick={successBtnHandler}>
-                                    승인
-                                </Button>
-                            </>
-                        )}
-                        {((clickedData?.status === "success" && role === "admin") ||
-                            (clickedData?.status === "waiting" && role === "worker")) && (
-                            <Button colorScheme="pink" onClick={deleteDocHandler}>
-                                삭제
-                            </Button>
-                        )}
-                    </ModalFooter>
-                </ModalContent>
-            </Modal>
-
-            <Modal isOpen={rejectModal.isOpen} onClose={rejectModal.onClose}>
-                <ModalOverlay />
-                <ModalContent>
-                    <ModalHeader>반려</ModalHeader>
-                    <ModalCloseButton />
-                    <Box as="form" onSubmit={rejectSubmitHandler}>
-                        <ModalBody>
-                            <FormControl isRequired>
-                                <FormLabel>사유</FormLabel>
-                                <Input
-                                    type="text"
-                                    placeholder="사유를 입력해주세요."
-                                    variant="flushed"
-                                    _placeholder={{ fontSize: "0.9rem" }}
-                                    ref={initialRef}
-                                    value={rejectReason}
-                                    onChange={changeRejectReason}
-                                />
-                            </FormControl>
-                        </ModalBody>
-
-                        <ModalFooter>
-                            <Button colorScheme="pink" mr={3} type="submit">
-                                반려
-                            </Button>
-                        </ModalFooter>
-                    </Box>
-                </ModalContent>
-            </Modal>
+            <DocDetailModal detailModal={detailModal} clickedData={clickedData} company={company} role={role} />
         </>
     );
 };
